+++ src/mimic/src/data/query/mod.rs
@@ -1,12 +1,10 @@
 mod delete;
 mod load;
 mod save;
-mod types;
 
 pub use delete::*;
 pub use load::*;
 pub use save::*;
-pub use types::*;
 
 use crate::{Error, deserialize, traits::EntityKind};
 use thiserror::Error as ThisError;